#362 refactor: domain.constant 패키지 내 파일들을 domain으로 이동
