+++ src/main/java/com/floney/floney/book/domain/ExcelDuration.java
@@ -1,4 +1,4 @@
-package com.floney.floney.book.domain.constant;
+package com.floney.floney.book.domain;
 
 import lombok.Getter;
