--- src/main/java/com/floney/floney/book/domain/constant/CategoryEnum.java
@@ -1,17 +1,0 @@
-package com.floney.floney.book.domain.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum CategoryEnum {
-
-    FLOW("내역"),
-    ASSET("자산"),
-    FLOW_LINE("내역분류");
-
-    private final String name;
-
-    CategoryEnum(final String name) {
-        this.name = name;
-    }
-}
