+++ src/main/java/com/floney/floney/book/service/ExcelServiceImpl.java
@@ -1,6 +1,6 @@
 package com.floney.floney.book.service;
 
-import com.floney.floney.book.domain.constant.ExcelDuration;
+import com.floney.floney.book.domain.ExcelDuration;
 import com.floney.floney.book.domain.entity.BookLine;
 import com.floney.floney.book.dto.request.ExcelDownloadRequest;
 import com.floney.floney.book.repository.BookLineRepository;
