+++ src/main/java/com/floney/floney/common/domain/vo/DateDuration.java
@@ -1,6 +1,6 @@
 package com.floney.floney.common.domain.vo;
 
-import com.floney.floney.book.domain.constant.ExcelDuration;
+import com.floney.floney.book.domain.ExcelDuration;
 import com.floney.floney.book.util.DateUtil;
 import com.floney.floney.common.exception.book.LimitRequestException;
 import lombok.Builder;
